requirements.txt
opencv-python
numpy
pycocotools
scikit-learn
scipy
torchvision
tqdm
json
six
transformers==4.23.0
re
pickle
random